feat: attempt result selector
Separated out submission form into its own component.
